Added a transform gizmo.
Grabbing game objects that don't have renderers is really difficult when the scene
is zoomed out by any amount, so this transform gizmo appears on an object in those
cases and lets you drag stuff around more easily. Also switched off picking when
the mouse is moving, because it was too slow on larger levels. If we miss it, it can
go back in.