File system changes #ADD:IFileSystemEventPattern interface
This interface can be implemented in plugins to allow Duality to recognize
sequences of file system events as the underlying user action that they
represent. For example, saving a file in Photoshop results in some temp
file creation, copies, renames, and deletes. Implementers of this interface
can present these actions to Duality as single events that more closely
represent the user action.